// FO/gestpat-frontend/src/app/models/private-pat.model.ts
export class PrivatePat {
  id?: string;
  registrationNumber?: string;
  type?: string;
  typeAr?: string;
  landReferences?: string;
  landReferencesAr?: string;
  registrationReference?: string;
  registrationDate?: Date;
  area?: number;
  location?: string;
  locationAr?: string;
  latitude?: number;
  longitude?: number;
  acquisitionSource?: string;
  acquisitionSourceAr?: string;
  purchasePrice?: number;
  zoningDesignation?: string;
  zoningDesignationAr?: string;
  currentUse?: string;
  currentUseAr?: string;
  marketValue?: number;
  isPrivatelyUsedByThirdParty?: boolean;
  privateUseDetails?: string;
  privateUseDetailsAr?: string;
  leaseAgreementDate?: Date;
  tenantName?: string;
  tenantNameAr?: string;
  leaseDuration?: number;
  rentalPrice?: number;
  increaseRatePercent?: number;
  rentPaymentDate?: Date;
  privateDomainRemovalReference?: string;
  privateDomainRemovalDate?: Date;
  privateDomainRemovalJustification?: string;
  privateDomainRemovalJustificationAr?: string;
  transferOrPublicDomainReference?: string;
  transferOrPublicDomainDate?: Date;
  transferPriceOrExchangeAmount?: number; 
  notes?: string;
  notesAr?: string;
}